Refactor: Change score type from Integer to BigDecimal to prevent type mismatch.

+++ src/main/java/com/group16/uno/dto/LeaderBoardDTO.java
@@ -1,11 +1,13 @@
 package com.group16.uno.dto;
 
+import java.math.BigDecimal;
+
 public class LeaderBoardDTO {
     private String username;
-    private Integer totalScore;
+    private BigDecimal totalScore;
 
 
-    public LeaderBoardDTO(String username, Integer totalScore) {
+    public LeaderBoardDTO(String username, BigDecimal totalScore) {
         this.username = username;
         this.totalScore = totalScore;
     }
@@ -19,11 +21,11 @@ public class LeaderBoardDTO {
         this.username = username;
     }
 
-    public Integer getTotalScore() {
+    public BigDecimal getTotalScore() {
         return totalScore;
     }
 
-    public void setTotalScore(Integer totalScore) {
+    public void setTotalScore(BigDecimal totalScore) {
         this.totalScore = totalScore;
     }
 }
